requirements.txt
google-adk
streamlit
mcp
python-dotenv
pydantic
